Add model.ValidateRelation function
To check if a given relation is valid according to the model
